pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "google-adk>=1.5.0",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.6",
    "langchain-openai>=0.3.27",
    "libmagic>=1.0",
    "litellm>=1.77.5",
    "python-socketio>=5.14.2",
    "requests>=2.32.4",
    "unstructured[md,pdf]>=0.18.2",
]
